requirements.txt
black==22.3.0
click==8.1.3
colorama==0.4.5
flake8==4.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.3.0
platformdirs==2.5.2
pycodestyle==2.8.0
pyflakes==2.4.0
pyocr==0.8.3
tomli==2.0.1
